refactor(webapp): merge budget-details models + data-core into view-models
Unify the two colocated folders into a single canonical `view-models/`.
Zero functional change - pure file relocation + import path updates.

Wave 1 Track A of the form-schemas normalization plan.
