+++ frontend/projects/webapp/src/app/feature/budget/budget-details/budget-grid/budget-grid-card.ts
@@ -16,7 +16,10 @@ import { FinancialKindDirective } from '@ui/financial-kind';
 import { FinancialKindIndicator } from '@ui/financial-kind-indicator';
 import { OriginalAmountLine } from '@ui/original-amount-line';
 import { RecurrenceLabelPipe } from '@ui/transaction-display';
-import { formatMatchAnnotation, type BudgetLineTableItem } from '../data-core';
+import {
+  formatMatchAnnotation,
+  type BudgetLineTableItem,
+} from '../view-models';
 import { SegmentedBudgetProgress } from '../components/segmented-budget-progress';
 import { BudgetActionMenu } from '../components/budget-action-menu';
