+++ frontend/projects/webapp/src/app/feature/budget/budget-details/budget-table/cells/balance-cell.ts
@@ -11,7 +11,7 @@ import { UserSettingsStore } from '@core/user-settings';
 import type {
   BudgetLineTableItem,
   TransactionTableItem,
-} from '../../data-core';
+} from '../../view-models';
 
 @Component({
   selector: 'pulpe-balance-cell',
