+++ frontend/projects/webapp/src/app/feature/budget/budget-details/budget-table/cells/name-cell.ts
@@ -15,7 +15,7 @@ import {
   formatMatchAnnotation,
   type BudgetLineTableItem,
   type TransactionTableItem,
-} from '../../data-core';
+} from '../../view-models';
 
 @Component({
   selector: 'pulpe-name-cell',
