+++ frontend/projects/webapp/src/app/feature/budget/budget-details/components/budget-action-menu.ts
@@ -13,7 +13,7 @@ import { MatTooltipModule } from '@angular/material/tooltip';
 import { TranslocoPipe } from '@jsverse/transloco';
 import type { BudgetLine, SupportedCurrency } from 'pulpe-shared';
 import { CURRENCY_CONFIG } from '@core/currency';
-import type { BudgetLineTableItem } from '../data-core';
+import type { BudgetLineTableItem } from '../view-models';
 
 const BALANCE_FORMATTERS = new Map<string, Intl.NumberFormat>();
