+++ frontend/projects/webapp/src/app/feature/budget/budget-details/view-models/budget-item-constants.ts
