+++ frontend/projects/webapp/src/app/feature/budget/budget-details/view-models/budget-item-data-builder.ts
@@ -13,7 +13,7 @@ import type {
   GroupHeaderTableItem,
   TableRowItem,
   TransactionTableItem,
-} from './budget-item-models';
+} from './table-items.view-model';
 import {
   KIND_ICONS,
   GROUP_LABELS,
